dom/promise/tests/test_promise_and_timeout_ordering_workers.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test for promise and timeout ordering in workers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
/* global async_test, assert_equals */
var t = async_test("Promise callbacks in workers should run immediately after the setTimeout handler that enqueues them");
var w = new Worker("file_promise_and_timeout_ordering.js");
w.onmessage = t.step_func_done(function(e) {
  assert_equals(e.data, "t1start, t1end, promise, t2");
});
</script>
